refactor(後端): 整理原始碼與測試目錄
