+++ apps/worker/src/connectors/cathaybk.ts
