+++ apps/worker/src/connectors/sinopac.ts
