+++ apps/worker/src/platform/access-auth.ts
