+++ apps/worker/src/platform/crypto.ts
