+++ apps/worker/src/routes/exchange-rates.ts
@@ -1,7 +1,7 @@
 import type { Hono } from "hono";
 import { z } from "zod";
-import type { AppBindings } from "../env";
-import { jsonError } from "../http";
+import type { AppBindings } from "../platform/env";
+import { jsonError } from "../platform/http";
 
 const updateSchema = z.object({
   rates: z.record(z.string().min(3).max(3), z.number().finite().positive())
